// apps/recipe/lib/api/generated/recipeAPI.ts
export type FoodType = (typeof FoodType)[keyof typeof FoodType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FoodType = {
	BREAKFAST: "BREAKFAST",
	LUNCH: "LUNCH",
	DINNER: "DINNER",
	DESSERT: "DESSERT",
	SNACK: "SNACK",
	DRINK: "DRINK",
} as const;
// ...
export type IngredientItemAmount = string | null;

export type IngredientItemUnit = string | null;

export interface IngredientItem {
	name: string;
	amount?: IngredientItemAmount;
	unit?: IngredientItemUnit;
}
// ...
export type RecipeReadFoodType = FoodType | null;

export type RecipeReadSimilarityScore = number | null;

export interface RecipeRead {
	name: string;
	description: string;
	ingredients: IngredientItem[];
	instructions: string;
	food_type?: RecipeReadFoodType;
	id: string;
	created_by: string;
	created_at: string;
	updated_at: string;
	like_count: number;
	favorite_count: number;
	status: RecipeStatus;
	is_generated: boolean;
	is_liked?: boolean;
	is_favorited?: boolean;
	similarity_score?: RecipeReadSimilarityScore;
}

export type RecipeStatus = (typeof RecipeStatus)[keyof typeof RecipeStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RecipeStatus = {
	PUBLISHED: "PUBLISHED",
	DRAFT: "DRAFT",
} as const;